started to refactor directory tree
