+++ package.json
@@ -7,11 +7,11 @@
     "build-dev": "rimraf ./dist && babel-node ./node_modules/webpack/bin/webpack --config ./webpack/webpack.config.js --progress --profile --colors",
     "clean": "rimraf ./dist && rimraf ./tmp",
     "deploy": "yarn && npm run build && npm run forever-stop || true && npm run forever-start",
-    "dev": "nodemon server.js --ignore app/ --ignore tests/ --exec babel-node",
-    "forever-start": "forever start --minUptime 1000ms --spinSleepTime 1000ms -a -l forever.log -o out.log -e err.log -c 'nodemon server.js --exitcrash --exec babel-node' server.js",
-    "forever-stop": "forever stop server.js",
-    "lint": "./node_modules/.bin/eslint app models routes server.js test webpack",
-    "start": "./node_modules/.bin/babel-node server",
+    "dev": "nodemon server/server.js --ignore app/ --ignore tests/ --exec babel-node",
+    "forever-start": "forever start --minUptime 1000ms --spinSleepTime 1000ms -a -l forever.log -o out.log -e err.log -c 'nodemon server/server.js --exitcrash --exec babel-node' server/server.js",
+    "forever-stop": "forever stop server/server.js",
+    "lint": "./node_modules/.bin/eslint app models routes server test webpack",
+    "start": "./node_modules/.bin/babel-node server/server.js",
     "test": "./node_modules/mocha/bin/mocha --compilers js:babel-core/register ./test/test.js"
   },
   "dependencies": {
