+++ routes/views/logout.js
